Got rid of duplicate code fragment
