+++ src/main/kotlin/at/bitfire/dav4jvm/property/caldav/SupportedCalendarData.kt
@@ -8,13 +8,9 @@ package at.bitfire.dav4jvm.property.caldav
 
 import at.bitfire.dav4jvm.Property
 import at.bitfire.dav4jvm.PropertyFactory
-import at.bitfire.dav4jvm.XmlUtils
+import at.bitfire.dav4jvm.readContentTypes
 import okhttp3.MediaType
-import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import org.xmlpull.v1.XmlPullParser
-import org.xmlpull.v1.XmlPullParserException
-import java.util.logging.Level
-import java.util.logging.Logger
 
 class SupportedCalendarData: Property {
 
@@ -43,19 +39,7 @@ class SupportedCalendarData: Property {
         override fun create(parser: XmlPullParser): SupportedCalendarData {
             val supported = SupportedCalendarData()
 
-            try {
-                XmlUtils.processTag(parser, CALENDAR_DATA_TYPE) {
-                    parser.getAttributeValue(null, CONTENT_TYPE)?.let { contentType ->
-                        var type = contentType
-                        parser.getAttributeValue(null, VERSION)?.let { version -> type += "; version=$version" }
-                        type.toMediaTypeOrNull()?.let { supported.types.add(it) }
-                    }
-                }
-            } catch(e: XmlPullParserException) {
-                val logger = Logger.getLogger(javaClass.name)
-                logger.log(Level.SEVERE, "Couldn't parse <resourcetype>", e)
-                return supported
-            }
+            readContentTypes(parser, CALENDAR_DATA_TYPE, supported.types::add)
 
             return supported
         }
